+++ src/accendino/__init__.py
@@ -1,2 +1,2 @@
 # Package
-__version__ = '0.5.9'
+__version__ = '0.5.10'