+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # _Accendino_ changelog
 
+## 0.5.10
+
+* updated accendino files for FreeBSD packages
+* rebuild an artifact when one of its dependency has been rebuilt more recently
+* included files can be included just once
+
+
 ## 0.5.9
 
 * added support for cross compilation